Implement solution to project euler 8 "largest product in series"
- Refactor "integers.Sum" to be in a "reducers" file
- Implement "integers.Product" (which calculates a product for a slice of ints)
- Implement the solution using sub-slices; b/c Go just uses a pointer with
a reference to the underlying array, and we're doing simple math loops, this
runs in > 5ms on my laptop on battery.
- I also implemented a more convoluted solution with a product of 13 digits,
and two pointers, one at either end, which divided out numbers from the head
and multiplied numbers to the tail -- but handling zero made this difficult.
Also, it didn't seem to have much better performance. But I don't care about
performance as much as I care about a clean, clear, modular solution